+++ samples/sampler/stories/default/VideoPlayer.js
@@ -149,6 +149,7 @@ export const _VideoPlayer = () => {
 				noMiniFeedback={boolean('noMiniFeedback', Config)}
 				noSlider={boolean('noSlider', Config)}
 				pauseAtEnd={boolean('pauseAtEnd', Config)}
+				playsInline={boolean('playsInline', Config, true)}
 				poster={poster}
 				seekDisabled={boolean('seekDisabled', Config)}
 				spotlightDisabled={boolean('spotlightDisabled', Config)}